+++ src/app/core/models/app-config.model.ts
@@ -1,0 +1,5 @@
+export interface AppConfigModel {
+    AppName: string,
+    Version: string,
+    ApiUrl: string
+}