+++ src/app/core/services/constant.token.ts
@@ -1,0 +1,10 @@
+import { InjectionToken } from '@angular/core';
+import { AppConfigModel } from '../models/app-config.model';
+
+export const constant = {
+    AppName: "AppName",
+    Version: "1.1",
+    ApiUrl: "http://..."
+}
+
+export const ConstantToken = new InjectionToken<AppConfigModel>('constant');