+++ src/app/core/services/generator.service.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@angular/core';
+import { generateNewId } from './gen-id.generator';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class GeneratorService {
+
+  constructor() { }
+
+  generate(n: number)
+  {
+    var result = '';
+    var characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    var charactersLength = characters.length;
+    for ( var i = 0; i < n; i++ ) {
+        result += characters.charAt(Math.floor(Math.random() * charactersLength));
+    }
+    return result;
+  }
+
+  getNewId = generateNewId;
+  
+}