+++ src/app/core/services/local-storage.service.ts
@@ -1,0 +1,11 @@
+import { InjectionToken } from "@angular/core";
+
+export class LocalStorage {
+    constructor() {}
+
+    addItemToStorage(item: string) {
+        localStorage.setItem(item, item);
+    }
+}
+
+export const LocalStorageToken = new InjectionToken<LocalStorage>('localStorage');